Make sure AntiSpoof mappings are mapping in the correct direction.
For example, 5 should map to S, not the other way around. Also
correcting some lowercase to uppercase mappings. There are a lot
more mappings in equivset.in that need to be fixed but this is a
good start. This should fix some problems with AbuseFilter.

Note that only equivset.in was directly edited, the other equivset
files were regenerated using the generateEquivset.php script.

Also adding a FIXME for characters being mapped to themselves.

WARNING: Don't try to view equivset.ser in gerrit unless you want
to lock up your browser.

Bug: 27987
Change-Id: I613f9917e8500d3606c15eacabb7862075aaeb6d