Make sure AntiSpoof mappings are mapping in the correct direction.
(This was previously merged, but reverted as we realized that we
weren't prepared to fix all of the filters that would be broken
due to the change. We now have a roll-out plan in place.)

For example, 5 should map to S, not the other way around. Also
correcting some lowercase to uppercase mappings. There are a lot
more mappings in equivset.in that need to be fixed but this is a
good start. This should fix some problems with AbuseFilter.

Note that only equivset.in was directly edited, the other equivset
files were regenerated using the generateEquivset.php script.

Also adding a FIXME for characters being mapped to themselves.

WARNING: Don't try to view equivset.ser in gerrit unless you want
to lock up your browser.

Note that updates to the parser tests in AbuseFilter must be merged
simultaneously.

Bug: 27987
Change-Id: Iccb3e50073bbbc2b979cb62dd0e129afd1c2e55f