Remove homemade implementation of strtr
To make use of AntiSpoof normalization on larger texts the homemade
implementation of strtr was replaced. In order to do this,
generateEquivset needed to be fixed to generate a set with no duplicates
and no empty characters as keys. Also, moved away from serializing the
set (equivset.ser) using codepoints, not needed anymore.

Bug: T172766
Change-Id: I731733671b650b6bb2f480c41c4f6f2d2f5c62e8