Add some missing scripts: N'Ko, Ol Chiki, Meetei Mayek, and Warang Citi
The comment for SCRIPT_NKO begins with NKo, as that is the name of the
Unicode block [1]; however, since the apostrophized N'Ko appears to be
more common, it’s included in parentheses.

Meetei Mayek Extensions is added, but commented out – this block only
contains characters for historical orthographies, so I assume it should
be treated like Old Persian, Deseret, etc. which are also commented out.

[1]: https://www.unicode.org/charts/PDF/U07C0.pdf

Bug: T246525
Change-Id: I98c321ed35c511698d6c66f0fbe92d6d7cc84509