Do not lookup conflicts for non-legal user names
While the extension itself calls SpoofUser::getConflicts only after
checking isLegal(), the CentralAuth extension does not. Protect the
function for the current and possible future usages by skipping the
lookup the conflict for non-legal user names.

Without this check the query ends in "su_normalized IS NULL" and returns
the first 5 non-legal usernames to show up as conflict.

Bug: T105626
Change-Id: Ifbb5d98a0438dc9b5311c4ebd00da2a48b686fbc