Use normal localisation message for username conflict error
Glueing together localisation messages using RawMessage results in
poor machine-readable output in the API.

For now, for compatibility with existing translations and on-wiki
overrides, we still glue them, but inside the normal message instead.
In the future the separate messages could be removed, and their text
moved to the new message.

Bug: T167163
Change-Id: Ie20408ec5d9ac0ca8259ee7ca115acf7ca23dba8